fix: allow geometric layers to have children
