+++ src/index.js
@@ -6,9 +6,9 @@ export { default as Marker } from './layers/markers/Marker.vue';
 export { default as vBounce } from './layers/markers/Bounceable.js';
 export { default as Popup } from './layers/Popup.vue';
 export { default as Tooltip } from './layers/Tooltip.js';
-export { default as Circle } from './layers/vector/Circle.js';
-export { default as Polyline } from './layers/vector/Polyline.js';
-export { default as Polygon } from './layers/vector/Polygon.js';
+export { default as Circle } from './layers/vector/Circle.vue';
+export { default as Polyline } from './layers/vector/Polyline.vue';
+export { default as Polygon } from './layers/vector/Polygon.vue';
 export { default as ZoomControl } from './controls/ZoomControl.js';
 export { default as ScaleControl } from './controls/ScaleControl.js';
 export { default as LocateControl } from './controls/LocateControl.js';
